fixup: remove the coins from the proof structure
